print.lineprof {lineprof}	R Documentation

Print reduced line profiling call stack.

Description

By default, whenever you print a line profile, it is collapsed to two levels deep. You can override this using the depth parameter. reduce_depth is the function that implements this behaviour - you may want to use it if you're processing the line profiling data in other ways.

Usage

## S3 method for class 'lineprof'
print(x, digits = 3, depth = 2, ...)

reduce_depth(x, depth = 2)

Arguments

x
a line profiling object

digits
number of digits to use for performance summaries

depth
depth of calls

...
other arguments passed on to format

Examples

# Line profiling the profile parsing code
x <- parse_prof(find_ex("read-delim.prof"))

x
print(x, depth = 3)
print(x, depth = Inf)

[Package lineprof version 0.1.9001 Index]